+++ doc/whats-new.rst
@@ -6,8 +6,12 @@ What's New
 v0.2.0 (unreleased)
 -------------------
 
-- Starfinder: added ``boost dX`` trait to :class:`Damage`
+Added support for Starfinder-specific mechanics:
 
+- Added ``boost dX`` trait to :class:`Damage`
+- Added ``primary_target`` parameter to :func:`check` and :func:`map_outcome`
+  to more easily model the Soldier's Primary Target ability.
+- Added two demo notebooks.
 
 
 v0.1.1 (2025-08-11)